Fix the SHOP2 tracer.
There were multiple bugs:
1. Matching items to be traced (as opposed to classes like :GOALS) was broken
because of a bug in the TRACE-PRINT macro.  A wayward quote meant that the
tracer was trying to match quoted expressions like (CAR GOAL1) instead of
matching the value of that expression.
2. Tracing specific kinds of item was broken simply because none of them were
being checked.  According to the docs, a user should have been able to do
something like (SHOP-TRACE (:GOAL FOO)) but, in fact, this would not work,
because it simply was never being checked.  I completed the partial support for
this kind of construct that was in the current version of SHOP2.