Fix problem with finding :INCLUDE files created by ASDF.
ASDF moves FASL files into a cache directory, which means that *load-truename*
doesn't point to a file in the directory with the source code, but instead
points to a file in the cache directory.  That was breaking the use of relative
pathnames in :INCLUDE in DEFDOMAIN.

Fixed this by adding to the DEFDOMAIN a cached copy of the filename that will be
cached at compile time (at macro-expansion time), instead of at load time when
that redirection has occurred.