Allow the user to ignore negation dependency problems.
We cannot, in general, compute the dependencies for a SHOP2 negation
formula (because it's potentially quantified).  This kludgy
modification raises a CERROR once, then allows the user to ignore all
further errors of this type.

This solution could be improved by adding a special error class for
all of these and then the programmer could handle them with CONTINUE.