Fix dependency tracking and negation causal links.
Previously, I was not building dependency links back to the initial state.
That was problematic for detecting failed tasks, because we wouldn't
have causal links for propositions that might be clobbered in a state
divergence.  That's now fixed.

Also fixed causal links for ground negations.  Quantified negations
are still problematic, since SHOP2 doesn't have a finite
domain of quantification.