Fixed issues in the plan tree.
While testing the minimal-subtree system, discovered substantial
issues in the new enhanced plan tree.  Specifically, primitive
actions in the plan sequence were not EQ to tasks in the plan
tree -- because of the way the plan sequence was populated.
A similar issue in method expansion caused variables not to be
bound properly in the plan tree.

To fix this, I needed to add an EXPANDED-TASK slot to the plan-tree
node structures, to hold the final task resulting from planning, where
the TASK slot holds pre-planning s-expressions.