Extensive improvements to PDDL treatment.
Changed the treatment of quantifiers (FORALL and EXISTS) in PDDL, so that
we can take advantage of the finite domain of quantification.

Found that IMPLY was not implemented correctly (i.e., not as described in
the manual).  Fixed.

Reorganized the test suite to integrate the tests for explicit stack search,
which previously caused unnecessary duplication of tests.