Substantial overhauls to the plan repair code.
1. The code for establishing the world state after a divergence was
wrong.  Going forward from the divergence, we must re-apply the world
state changes as far as we can.  "As far as we can," means "until the
earliest action whose preconditions are violated."  Previously, I was
rolling forward all the way to the end of the now-broken plan, and
garbling the state contents.
2. Make sure that *DOMAIN* is bound properly when replanning.  There
was another bug with it having the wrong variable.