Multiple fixes to dependency handling.
- Previously assumed that dependencies were handled the same as bindings --
multiple responses that don't change the binding lists were treated as no-ops
(dependencies not collected).  This was OK for when we are doing all solutions
(disjunction), because multiple dependency sets for the same bindings are
redundant justifications. But in a /conjunctive/ setting, dependencies should
accrue, even if the binding lists don't change.

In the process, did some code reformatting.