Optimize special case of :FIRST goals.
Treat singleton :FIRST goals specially.  By doing this, we can save extra
computations -- if a :FIRST goal is an (implicit) conjunction of more than one
goal, then we must find all bindings of the subgoals to find satisfying bindings
for the conjunction.  But if it's a single literal, we need only find one
solution.