Remove redundant adds-then-deletes.
There are a number of actions that first add and then delete AVAILABLE and
CHANNEL_FREE in the same effect specification.
I suspect that these are due to a lazy translation of a temporal action, but I
am cutting them out, although they are theoretically correct, because they trip
some of my error-checking in plan repair.