Reorganize the handling of distinguished symbols.
Previously, distinguished symbols for variables, anonymous variables, and
primitive operators, were handled by setting a property on the property list.
This was efficient, but required programmers to remember to invoke the function
SET-VARIABLE-PROPERTY on every piece of SHOP2 code to assure that symbols were
properly categorized.

This change makes the call to SET-VARIABLE-PROPERTY optional (it could
eventually be removed) by making the property list setting be done in a
memoizing style.  That is, if the property is not set on a symbol, we set it
(the same way that SET-VARIABLE-PROPERTY would), but if it is set, we use that.