Clarifying comment.
Plan repair does some odd things with the tags (integers) in the SHOP3
world states.  They were originally just integers in order, but in
order to do the plan repair, I needed to be able to inject a new
state (for the divergences) into the middle of a sequence.  So now
they are sequences of *even* integers by default, and the extra one I
add for plan divergences is an odd-numbered one between two previously
consecutive states.