Update tests.
Update tests for reformulated version of PDDL type checks.
Also fix bugs in test data.

Some of the test data were using a non-standard extension to the reader
that permits an entire s-expression, instead of just a symbol, to be
package-qualified. Discovered that was non-standard when it broke
Lispworks.

In the process of fixing this for Lispworks, ended up reformatting the
file.