Merge clean-ups from deprecated-implicit-conjunction (#85)
Originally, I had thought it made sense to remove all use of implicit conjunction (accepting lists of literals as well as `(and ...)`).  But it caused a fair amount of disruption to the tests, and in discussions with @ukuter it seemed better to simply remove implicit conjunction in new syntax, like our new `:op` syntax for operators, and an upcoming new syntax for methods.

However, there were a number of clean-ups done in passing, so I pulled them out of the `deprecated-implicit-conjunction` branch so they could be saved.  They appear in this merge.

Fixes include:
* cleaning up the processing of preconditions in domain definitions.  The same function, `process-pre` was being used to process both preconditions *and* add and delete lists.  This was bad because the two have rather different syntax. So they have been pulled apart.
* Fixed up plan validation and added a number of validation checks to the tests.
* There are also some miscellaneous fixes, like making the include directives in the openstacks domain definitions more robust.