Revise explicit stack return collection.
The explicit state search was previously only able to collect a single result plan. This is a partial step towards allowing it to find multiple plans.
This commit introduces a new data structure that collects all the values that need to be returned when a single plan is found, but still only returns one plan.
This will later be expanded to being able to collect multiple plans.