Fix primitive task handling in tree extraction.
Previously, we errored if we had a record of a primitive task for
which we didn't have a corresponding action in the plan.  However,
that was wrong because we may have nodes in our recorded reductions
that never produced plans.  So instead of raising an error, we ignore
these entries.