Shadow CL:CHECK-TYPE.
I was using `check-type` because it was a handy shorthand, but it
turns out to mess up SBCL's type checking, because it's got a restart
where the user can reset the checked place.

Instead, replace with a macro that just checks the type and raises a
type error if the object is not correctly typed.