Extensions to HDDL plan translation.
The HDDL domains that are generated in plan repair for rewrite
sometimes violate the normal constraint that all nodes be ground.

Note: This should probably be fixed, but might happen only at the cost
of making the resulting domains much less efficient to plan.

Changes here allow this normal constraint to be overridden, but note
that if we do, the resulting HDDL plans might count as ill-formed to a
verifier.

Also add a new `find-all-tree-nodes-if` to the PLAN-TREE (enhanced
plan tree) package.